+++ src/components/LoQEExplorer.vue
@@ -446,6 +446,9 @@
                     <pre class="text-caption mt-1" style="white-space: pre-wrap">{{
                       activeError
                     }}</pre>
+                    <template v-if="activeError.includes('CORS') || activeError.includes('Failed to fetch') || activeError.includes('HTTP Error')" #append>
+                      <CorsConfigDialog />
+                    </template>
                   </v-alert>
 
                   <!-- Result Tabs (only shown when multiple statements) -->
@@ -702,6 +705,7 @@ import type { LoQEQueryResult } from '../composables/loqe/types';
 import SqlEditor from './SqlEditor.vue';
 import LoQENavigationTree from './LoQENavigationTree.vue';
 import DuckDBSettingsDialog from './DuckDBSettingsDialog.vue';
+import CorsConfigDialog from './CorsConfigDialog.vue';
 import { helix, hourglass } from 'ldrs';
 
 helix.register();
